requirements.txt
sentence-transformers==3.3.1
python-telegram-bot==21.9
qdrant-client==1.12.2
dotenv-python==0.0.1
FlashRank==0.2.9
ollama==0.4.4
numpy==2.2.0